pyproject.toml
[tool.poetry]
name = "distributed-task-scheduler"
version = "0.1.0"
description = ""
authors = ["Shiyan Shirani <shiyan99s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
kombu = "^5.4.2"
flower = "^2.0.1"
gunicorn = "^23.0.0"
celery = "^5.4.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.7.3"
black = "^24.10.0"

[tool.ruff]
fix-only = true
line-length = 88
select = ["E", "F", "W", "C90"]
ignore = ["N803", "N802"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
